third_party/blink/web_tests/platform/generic/fast/events/page-visibility-bubble-expected.txt

This test checks that 'visibilitychange' event bubbles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.visibilityState is "visible"
PASS document.hidden is false
PASS event.bubbles is true
PASS event.bubbles is true
PASS document.visibilityState is "hidden"
PASS document.hidden is true
PASS event.bubbles is true
PASS event.bubbles is true
PASS successfullyParsed is true

TEST COMPLETE
